Add solution for Day 05
this is working but broken. I'm doing something dumb that's causing
the output from the first part of the puzzle to be re-used as the
input to the second part of the puzzle. Likely because I'm not making
a deep copy of the input list of crates. I'll fix in the next commit
but making coffee first...